Se agrega imagen en menu, se agrega como recurso del proyecto y se valida que no se pueda eliminar beneficiario con psicologo

+++ WindowsFormsApplication1/WindowsFormsApplication1/F_Beneficiario.cs
@@ -42,6 +42,7 @@ namespace WindowsFormsApplication1
         {
             // TODO: esta línea de código carga datos en la tabla 'marillacDataSet.Beneficiario' Puede moverla o quitarla según sea necesario.
             this.beneficiarioTableAdapter.Fill(this.marillacDataSet.Beneficiario);
+            limpiarTextBox();
         }
 
         private void botonSalir_Click(object sender, EventArgs e)
@@ -58,10 +59,11 @@ namespace WindowsFormsApplication1
                 string query = "INSERT INTO Persona.Beneficiario(nombre, paterno, materno, direccion, colonia, telefono, sexo, fechaNacimiento) VALUES('" + nombre.Text + "', '" + paterno.Text + "', '" + materno.Text + "', '" + direccion.Text + "', '" + colonia.Text + "'," + telefono.Text + ", '" + sexo.Text + "', '" + dateTimePicker1.Value.Year + "-" + dateTimePicker1.Value.Month + "-" + dateTimePicker1.Value.Day + "')";
                 adapter.InsertCommand = new SqlCommand(query, con);
                 adapter.InsertCommand.ExecuteNonQuery();
-                limpiarTextBox();
+               
                 con.Close();
 
                 this.beneficiarioTableAdapter.Fill(this.marillacDataSet.Beneficiario);
+                limpiarTextBox();
 
             }
             catch (Exception ex)
@@ -80,31 +82,42 @@ namespace WindowsFormsApplication1
             colonia.Clear();
             telefono.Clear();
             sexo.Clear();
+            psicologo.Clear();
             dateTimePicker1.Value = new DateTime(1990, 01, 01);
         }
 
         private void botonEliminar_Click(object sender, EventArgs e)
         {
             Int64 idEliminar;
+            Int64 idPsicologo;
 
             if (dataGridView1.SelectedRows.Count != 0)
             {
                 idEliminar = Convert.ToInt64(dataGridView1.Rows[dataGridView1.CurrentCell.RowIndex].Cells[0].Value.ToString());
-                try
+                try { idPsicologo = Convert.ToInt64(dataGridView1.Rows[dataGridView1.CurrentCell.RowIndex].Cells[4].Value.ToString()); }
+                catch { idPsicologo = -1; }
+                if(idPsicologo<0)
                 {
-                    con.Open();
-                    string query = "DELETE FROM Persona.Beneficiario WHERE idBeneficiario=" + idEliminar;
-                    adapter.InsertCommand = new SqlCommand(query, con);
-                    adapter.InsertCommand.ExecuteNonQuery();
-                    con.Close();
+                    try
+                    {
+                        con.Open();
+                        string query = "DELETE FROM Persona.Beneficiario WHERE idBeneficiario=" + idEliminar;
+                        adapter.InsertCommand = new SqlCommand(query, con);
+                        adapter.InsertCommand.ExecuteNonQuery();
+                        con.Close();
 
-                    this.beneficiarioTableAdapter.Fill(this.marillacDataSet.Beneficiario);
-                    limpiarTextBox();
+                        this.beneficiarioTableAdapter.Fill(this.marillacDataSet.Beneficiario);
+                        limpiarTextBox();
+                    }
+                    catch (Exception ex)
+                    {
+                        con.Close();
+                        MessageBox.Show(ex.Message.ToString());
+                    }
                 }
-                catch (Exception ex)
+                else
                 {
-                    con.Close();
-                    MessageBox.Show(ex.Message.ToString());
+                    MessageBox.Show("No se puede eliminar a un beneficiario que tenga un psicólogo asignado");
                 }
             }
         }
